+++ CHANGES.md
@@ -3,6 +3,17 @@ PyScript Changelog
 
 Here you can see the full list of changes between each PyScript release.
 
+Version 0.2.5
+-----------
+- added: `str.split` now mimicks python string split
+- changed: `requests.interceptors` must return explicitly false to exit handling of response
+- changed: `requests.interceptors` now takes an object which can be for requests as well
+- changed: renamed `list.invoke` with `list.map`
+- fixed: `requests.upload` can now specify headers
+- fixed: `cache.storeFile` should now work with `File` objects
+- removed: `str.sprintf`, which wasn't working in previous version
+- removed: `requests.beforeRequest`, replaced by `requests.interceptors`
+
 Version 0.2.4
 -----------
 - deprecated: `str.sprintf` deprecated in favor of `str.format`